backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic[email]==2.10.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
fastapi-mail==1.4.1
python-multipart==0.0.6
aiofiles==23.2.1
twilio==9.3.0  # For SMS/WhatsApp
# For rate limiting
slowapi==0.1.9
redis==5.0.7

# For mobile money (example - adapt based on provider)
requests==2.32.3
cryptography==43.0.3

# For push notifications (Firebase example)
pyfcm==1.5.4

# For document verification
pillow==11.0.0
python-multipart==0.0.20